+++ plone/app/blocks/tests/test_layoutbehavior.py
@@ -1,10 +1,12 @@
 # -*- coding: utf-8 -*-
+from plone.app.blocks.layoutbehavior import ILayoutAware
+from plone.app.blocks.layoutbehavior import ILayoutBehaviorAdaptable
+from plone.app.blocks.layoutbehavior import LayoutAwareTileDataStorage
 from plone.app.blocks.testing import BLOCKS_FUNCTIONAL_TESTING
 from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
 from plone.registry.interfaces import IRegistry
 from plone.uuid.interfaces import IUUID
-from zope.component import getGlobalSiteManager
 from zope.component import getUtility
 
 import pkg_resources
@@ -28,37 +30,27 @@ class TestLayoutBehavior(unittest.TestCase):
         self.request = self.layer['request']
         self.registry = getUtility(IRegistry)
 
+        # Enable layout aware behavior
+        fti = self.portal.portal_types['Document']
+        behaviors = [i for i in fti.behaviors]
+        behaviors.extend([
+            'plone.app.blocks.layoutbehavior.ILayoutAware'
+        ])
+        behaviors = tuple(set(behaviors))
+        fti._updateProperty('behaviors', behaviors)
+
         setRoles(self.portal, TEST_USER_ID, ('Manager',))
         self.portal.invokeFactory('Folder', 'f1', title=u"Folder 1")
         self.portal['f1'].invokeFactory('Document', 'd1', title=u"Document 1")
         setRoles(self.portal, TEST_USER_ID, ('Member',))
 
-        if HAS_PLONE_APP_CONTENTTYPES:
-            from plone.app.contenttypes.interfaces import IDocument
-            iface = IDocument
-        else:
-            iface = self.portal['f1']['d1'].__class__
-
-        from plone.app.blocks.layoutbehavior import ILayoutAware
-        from plone.app.blocks.layoutbehavior import LayoutAwareBehavior
-
-        sm = getGlobalSiteManager()
-        sm.registerAdapter(LayoutAwareBehavior, [iface])
-        registrations = sm.getAdapters(
-            (self.portal['f1']['d1'],),
-            ILayoutAware
-        )
-        self.assertEqual(len(list(registrations)), 1)
         self.behavior = ILayoutAware(self.portal['f1']['d1'])
-        registrations = sm.getAdapters(
-            (self.portal['f1']['d1'],),
-            ILayoutAware
-        )
-        self.assertEqual(len(list(registrations)), 1)
+        self.assertTrue(
+            ILayoutBehaviorAdaptable.providedBy(self.behavior.context))
 
-    def test_content(self):
+    def test_custom_content_layout(self):
         from plone.app.blocks.layoutviews import ContentLayoutView
-        self.behavior.content = \
+        self.behavior.customContentLayout = \
             u'<html><body><a href="{0:s}"></a></body></html>'.format(
                 'resolveuid/{0:s}'.format(IUUID(self.portal['f1'])))
         rendered = ContentLayoutView(self.portal['f1']['d1'], self.request)()
@@ -66,3 +58,37 @@ class TestLayoutBehavior(unittest.TestCase):
         # Test that UUID is resolved by outputfilters
         self.assertNotIn(IUUID(self.portal['f1']), rendered)
         self.assertIn(self.portal['f1'].absolute_url(), rendered)
+
+    def test_content(self):
+        tile = 'plone.app.tiles.demo.transient/demo'
+        self.behavior.content = u"""\
+<html>
+<body>
+<div data-tile="@@plone.app.tiles.demo.transient/demo"
+data-tiledata='{"message": "Hello World!"}' />
+</body>
+</html>
+"""
+        view = self.portal['f1']['d1'].restrictedTraverse(tile)()
+        self.assertEqual(
+            view,
+            '<html><body><b>Transient tile Hello World!</b></body></html>'
+        )
+
+        storage = LayoutAwareTileDataStorage(self.portal['f1']['d1'],
+                                             self.layer['request'])
+        self.assertEqual(len(storage), 1)
+        self.assertEqual(list(storage), [tile])
+        self.assertEqual(storage[tile]['message'], u"Hello World!")
+
+        data = storage[tile]
+        data['message'] = u"Foo bar!"
+        storage[tile] = data
+
+        delattr(self.layer['request'], '__annotations__')  # purge memoize
+
+        view = self.portal['f1']['d1'].restrictedTraverse(tile)()
+        self.assertEqual(
+            view,
+            '<html><body><b>Transient tile Foo bar!</b></body></html>'
+        )