+++ CHANGES.rst
@@ -6,8 +6,16 @@ Changelog
 
 Incompatibilities:
 
+- Remove official Plone 4 support
+  [datakurre]
+
 - Removed pluggable grid framework
   [agitator]
+
+- Moved ``ILayoutAware.content`` to ``ILayoutAware.customContentLayout``.
+  Added ``ILayoutAware.content`` as layout independent, but layout
+  like, tile data storage
+  [datakurre]
 
 - Moved functions ``getDefaultAjaxLayout``, ``getDefaultSiteLayout``, ``getLayout`` and ``getLayoutAwareSiteLayout`` to ``.layoutbehavior`` in order to avoid circular imports
   (all deprecated now, see section New).